tg_bot/requirements.txt
python-telegram-bot
random-word
selenium
requests
webdriver_manager
pillow
python-io